Update to toolbox
the toolbox now:
- automatically detects the OS of the user to pic the appropriate script
- automatically loads the latest simulation result after init (if one exists) and after new simulation runs
- run_simulation() does work within and without jupyter notebooks
- initialization does not change the working directory of the process any longer, this would for example make it impossible to have two instances of the toolbox at the same time. All paths are relative to the provided "mosaic_path" now
- some other minor changes

In general I think we need to further improve the API, currently users can access a lot of useless data and it's still hard to actually identify what the interesting results are. For this of course we first need to define what parameters they are supposed to change and following, what results get affected by that. A simple idea for a new API endpoint would be e.g. "23 cars went left, 77 cars went right", so they can directly see the impact of their changes